Task 4 - Accidentally did task 4b as well when bug testing, so decided to bundle the tasks 4a and 4b together.
Created a rotating axe that follows the players current position at all times, and despawns at a certain time. Axes randomly spawn in the range set, including unfair spawns that will hit the player immediately.

Created a spawner that instatiates the spinning axes and a button that spawns axes seperate from the timer.